% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_optim.R
\name{admm_optim}
\alias{admm_optim}
\title{ADMM Algorithm for Meta-analysis/Multiple Outcomes Personalized Medicine}
\usage{
admm_optim(x, y, p, q, lambda1, lambda2, lambda3, abs.tol = 1e-05,
  rel.tol = 1e-05, maxit = 500L, rho = NULL)
}
