% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mp}
\alias{coef.mp}
\title{Coefficients of a Fitted "mp" Object}
\usage{
\method{coef}{mp}(object, ...)
}
\arguments{
\item{object}{A fitted "mp" object returned by "mpersonalized".}

\item{...}{not used}
}
\value{
A list object. Each element in the list is the fitted coefficients
corresponding to one penalty parameter value in \code{mp$penalty_parameter_sequence}.
}
\description{
This function provides \code{coef} method for "mp" class objects.
}
\examples{
set.seed(123)
sim_dat  = simulated_dataset(n = 200, problem = "meta-analysis")
Xlist = sim_dat$Xlist; Ylist = sim_dat$Ylist; Trtlist = sim_dat$Trtlist

# fit different rules with SGL penalty for this meta-analysis problem
mp_mod_diff = mpersonalized(problem = "meta-analysis",
                            Xlist = Xlist, Ylist = Ylist, Trtlist = Trtlist,
                            penalty = "SGL", single_rule = FALSE)

mp_coef = coef(mp_mod_diff)
set.seed(NULL)
}
