% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mp_cv}
\alias{coef.mp_cv}
\title{Coefficients of a Fitted "mp_cv" Object}
\usage{
\method{coef}{mp_cv}(object, ...)
}
\arguments{
\item{object}{A fitted "mp" object returned by "mpersonalized".}

\item{...}{not used}
}
\value{
The fitted coefficients corresponding to the optimal penalty parameter.
}
\description{
This function provides \code{coef} method for "mp_cv" class objects.
}
\examples{
set.seed(123)
sim_dat  = simulated_dataset(n = 200, problem = "meta-analysis")
Xlist = sim_dat$Xlist; Ylist = sim_dat$Ylist; Trtlist = sim_dat$Trtlist

# fit different rules with lasso penalty for this meta-analysis problem
mp_cvmod_diff = mpersonalized_cv(problem = "meta-analysis",
                                 Xlist = Xlist, Ylist = Ylist, Trtlist = Trtlist,
                                 penalty = "lasso", single_rule = FALSE)

mp_coef = coef(mp_cvmod_diff)
set.seed(NULL)
}
