% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_estimate.R
\name{lambda_estimate}
\alias{lambda_estimate}
\title{Default Penalty Parameter Sequence if Not Given}
\usage{
lambda_estimate(modelXlist, modelYlist, penalty, single_rule, alpha,
  num_lambda1, num_lambda2, num_single_rule_lambda)
}
\arguments{
\item{modelXlist}{the Xlist in contrast framework after standardization}

\item{modelYlist}{the Ylist in contrast framework after standardization}

\item{penalty}{penalty type}

\item{single_rule}{whether a single treatment rule is required}
}
\value{
estimated lambda for required penalty if not provided
}
