% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPersonalized.R
\name{MetaPersonalized}
\alias{MetaPersonalized}
\title{Meta-analysis/Multiple Outcomes for Personalized Medicine}
\usage{
MetaPersonalized(Xlist, Ylist, Trtlist, Plist, typelist = NULL,
  model = c("linear", "lasso", "GL", "SGL", "fused", "lasso+fused",
  "GL+fused", "SGL+fused"), lambda1 = NULL, lambda2 = NULL,
  unique_rule_lambda = NULL, alpha = NULL, unique_rule = FALSE)
}
\arguments{
\item{Xlist}{a list object with \eqn{k}th element denoting the covariate matrix of study k}

\item{Ylist}{a list object with \eqn{k}th element denoting the response vector of study k}

\item{Trtlist}{a list object with \eqn{k}th element denoting the treatment vector of study k (coded as 0 or 1)}

\item{model}{the model to be used for the above framework, can be linear, meta-analysis,
sparse group lasso, group lasso or lasso}

\item{lambda1}{lambda1 in the framework above}

\item{lambda2}{lambda2 in the framework above}

\item{unique_rule_lambda}{\eqn{\lambda_{uni}} when unique treatment rule is required}

\item{alpha}{alpha in the framework above}

\item{unique_rule}{a logical value, whether a unique treatment rule is required}

\item{typlelist}{a list object with \eqn{k}th element denoting the type of response in study k, can be continuous or binary, default is continuous}
}
\value{
an S3 object of class "mp", which contains the information of the fitted model. It could be supplied
to the predict function
}
\details{
Assume the total number of studies is K. This function implements meta-analysis for personalized medicine based on the following framework:
\deqn{ \min_{g_1,\dots,g_K} \sum_{k=1}^K \sum_{i=1}^{n_k}\frac{|\hat{C}_k(X_{i})|}{\sum_{i=1}^{n_k}|\hat{C}_k(X_{i})|}\bigl [1\{\hat{C}_k(X_{i})>0\}-g_k(X_{i})\bigr]^2 + h(g_1,\dots,g_K)}
Here the regularization function \eqn{h} is of the form of a sum of sparse group lasso and fused lasso penalty
\deqn{h = (1-\alpha)\lambda_1\sqrt{q} \sum_{j=1}^p \|\boldsymbol{\beta_j}\|_2+\alpha \lambda_1  \sum_{j=1}^p \|\boldsymbol{\beta_j}\|_1+ \lambda_2 \sum_{j=1}^p \sum_{1\le a < b \le K}|\beta_{ja}-\beta_{jb}|}
where \eqn{\boldsymbol{\beta_j}=(\beta_{j1},\dots,\beta_{jK})}

If we would like a unique rule to be obtained, we should let \eqn{g_1 = \dots= g_K} and
\deqn{h = \lambda_{uni} \|\beta\|_1}

By setting \eqn{\lambda_1, \lambda_2, \alpha} differently, different models could be obtained.
\itemize{
\item If \eqn{\lambda_1,\lambda2 \ne 0} and \eqn{alpha \ne 0} or \eqn{1}, it is sparse group lasso + fused model.
\item If \eqn{\lambda_2 = 0} and \eqn{\alpha \ne 0} or \eqn{1}, a sparse group lasso model is fitted.
\item If \eqn{\lambda_2 = 0} and \eqn{\alpha = 0}, a group lasso model is fitted.
\item If \eqn{\lambda_2 = 0} and \eqn{\alpha = 1}, a lasso model is fitted.
\item If \eqn{\lambda_1, \lambda_2 = 0}, a linear model is fitted.
}
}
