% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPersonalized_cv.R
\name{MetaPersonalized_cv}
\alias{MetaPersonalized_cv}
\title{Meta-analysis/Multiple Outcomes for Personalized Medicine with Cross Validation}
\usage{
MetaPersonalized_cv(Xlist, Ylist, Trtlist, Plist, typelist = NULL,
  model = c("lasso", "GL", "SGL", "fused", "lasso+fused", "GL+fused",
  "SGL+fused"), lambda1 = NULL, lambda2 = NULL, unique_rule_lambda = NULL,
  alpha = NULL, unique_rule = FALSE, cv_folds = 5)
}
\arguments{
\item{Xlist}{a list object with \eqn{k}th element denoting the covariate matrix of study k}

\item{Ylist}{a list object with \eqn{k}th element denoting the response vector of study k}

\item{Trtlist}{a list object with \eqn{k}th element denoting the treatment vector of study k (coded as 0 or 1)}

\item{model}{the model to be used for the above framework, can be meta-analysis,
sparse group lasso, group lasso or lasso(linear does not need tuning)}

\item{lambda1}{lambda1 in the framework above}

\item{lambda2}{lambda2 in the framework above}

\item{unique_rule_lambda}{\eqn{\lambda_{uni}} when unique treatment rule is required}

\item{alpha}{alpha in the framework above}

\item{unique_rule}{a logical value, whether a unique treatment rule is required}

\item{cv_folds}{number of folds needed for cross-validation, default is 5}

\item{typlelist}{a list object with \eqn{k}th element denoting the type of response in study k, can be continuous or binary, default is continuous}
}
\value{
an S3 object of class "mp_cv", which contains the information of the model with the best fitted lambda. It can be supplied to the predict function.
}
\description{
Meta-analysis/Multiple Outcomes for Personalized Medicine with Cross Validation
}
