% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mersonalized.R
\name{mpersonalized}
\alias{mpersonalized}
\title{Meta-analysis/Multiple Outcomes for Personalized Medicine}
\usage{
mpersonalized(problem = c("meta-analysis", "multiple outcomes"), X, Trt, P,
  Xlist, Ylist, Trtlist, Plist, typelist = NULL, penalty = c("none",
  "lasso", "GL", "SGL", "fused", "lasso+fused", "GL+fused", "SGL+fused"),
  lambda1 = NULL, lambda2 = NULL, unique_rule_lambda = NULL,
  alpha = NULL, unique_rule = FALSE)
}
\arguments{
\item{problem}{a character specifiy whether you want to solve "meta-analysis" or "multiple outcomes" problem. For "meta-analysis" problem,
the user should supply \code{Xlist}, \code{Ylist}, \code{Trtlist} and \code{Plist}. For "multiple outcomes" problem,
the user should supply \code{X}, \code{Ylist}, \code{Trt} and \code{P}.}

\item{X}{the covariate matrix that should be supplied when the problem is "multiple outcomes" with rows indicating subjects and columns indicating covariates.}

\item{Trt}{the treatment vector that should be supplied when the problem is "multiple outcomes". It should be coded as 0 or 1.}

\item{P}{the propensity score vector when the problem is "multiple outcomes".}

\item{Xlist}{a list object with \eqn{k}th element denoting the covariate matrix of study \eqn{k}. This should be supplied when the problem is
"meta-analysis".}

\item{Ylist}{When the problem is "meta-analysis", \code{Ylist} should be a list object with \eqn{k}th element denoting the response vector of study \eqn{k}. When the
problem is "multiple outcomes", \code{Ylist} should be a list object with \eqn{k}th element denoting the \eqn{k}th outcome.}

\item{Trtlist}{a list object with \eqn{k}th element denoting the treatment vector of study \eqn{k} (coded as 0 or 1). This should be supplied when the problem is
"meta-analysis".}

\item{Plist}{a list object with \eqn{k}the element denoting the propensity score vector of study \eqn{k}.}

\item{penalty}{For different rules, the penalty could be "none", "lasso", "GL", "SGL", "fused",
"lasso+fused", "GL+fused", "SGL+fused". For unique rule, the penalty could be "none" or "lasso".}

\item{lambda1}{lambda1 supplied in the framework when different rules are used.}

\item{lambda2}{lambda2 supplied in the framework when different rules are used.}

\item{unique_rule_lambda}{\eqn{\lambda} when unique rule is used.}

\item{alpha}{alpha in the framework when different rules are used.}

\item{unique_rule}{a logical value, whether a unique treatment rule is required}

\item{typlelist}{a list object with \eqn{k}th element denoting the type of response corresponding to the \eqn{k}th element in the list \code{Ylist}.
Each element should be "continuous" or "binary".}
}
\value{
an S3 object of class "mp", which contains the information of the fitted model. It could be supplied
to the predict function
}
\details{
Assume the total number of studies is \eqn{K}. This function is aimed to solve meta-analysis/multiple outcomes problems for personalized medicine based on the following framework:
\deqn{ \min_{g_1,\dots,g_K} \frac{1}{2}\sum_{k=1}^K \sum_{i=1}^{n_k}\frac{|\hat{C}_k(X_{i})|}{\sum_{i=1}^{n_k}|\hat{C}_k(X_{i})|}\bigl [1\{\hat{C}_k(X_{i})>0\}-g_k(X_{i})\bigr]^2 + h(g_1,\dots,g_K)}
Here the regularization function \eqn{h} is of the form of a sum of sparse group lasso and fused lasso penalty
\deqn{h = (1-\alpha)\lambda_1\sqrt{q} \sum_{j=1}^p \|\boldsymbol{\beta_j}\|_2+\alpha \lambda_1  \sum_{j=1}^p \|\boldsymbol{\beta_j}\|_1+ \lambda_2 \sum_{j=1}^p \sum_{1\le a < b \le K}|\beta_{ja}-\beta_{jb}|}
where \eqn{\boldsymbol{\beta_j}=(\beta_{j1},\dots,\beta_{jK})}

If we would like a unique rule to be obtained, we let \eqn{g_1 = \dots= g_K} and solve the following question instead
\deqn{\min_{g} \frac{1}{2}\sum_{k=1}^K \sum_{i=1}^{n_k}\frac{|\hat{C}_k(X_{i})|}{\sum_{i=1}^{n_k}|\hat{C}_k(X_{i})|}\bigl [1\{\hat{C}_k(X_{i})>0\}-g(X_{i})\bigr]^2 + h(g_1,\dots,g_K) + \lambda_{uni} \|\beta\|_1}

If we want different rules, by setting \eqn{\lambda_1, \lambda_2, \alpha} differently, different penalties can be obtained.
\itemize{
\item If \eqn{\lambda_1, \lambda_2 \ne 0} and \eqn{\alpha \ne 0} or \eqn{1}, the penalty is "SGL+fused".
\item If \eqn{\lambda_1, \lambda_2 \ne 0} and \eqn{\alpha = 0}, the penalty is "GL+fused".
\item If \eqn{\lambda_1, \lambda_2 \ne 0} and \eqn{\alpha = 1}, the penalty is "lasso+fused".
\item If \eqn{\lambda_1 = 0, \lambda_2 \ne 0}, the penalty is "fused".
\item If \eqn{\lambda_1 \ne0, \lambda_2 = 0} and \eqn{\alpha \ne 0} or \eqn{1}, the penalty is "SGL".
\item If \eqn{\lambda_1 \ne0, \lambda_2 = 0} and \eqn{\alpha = 0}, the penalty is "GL".
\item If \eqn{\lambda_1 \ne0, \lambda_2 = 0} and \eqn{\alpha = 1}, the penalty is "lasso".
\item If \eqn{\lambda_1, \lambda_2 = 0}, there is no penalty.
}

If we want unique rule,
\itemize{
\item If \eqn{\lambda \ne 0}, the penalty is "lasso".
\item If \eqn{\lambda = 0}, there is no penalty.
}
}
