% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mp}
\alias{plot.mp}
\title{Plot for a 'mp' Class Object}
\usage{
\method{plot}{mp}(mp, ind1, ind2, unique_ind)
}
\arguments{
\item{mp}{the 'mp' class object returned by MetaPersonalzied function}

\item{ind1}{the index of the wanted lambda1}

\item{ind2}{the index of the wanted lambda2}

\item{unique_ind}{the index of the wanted unique_rule_lambda}
}
\value{
a plot that
}
