% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCVE}
\alias{plotCVE}
\title{Cross Validation Error Plot for an "mp_cv" Class Object.}
\usage{
plotCVE(mp_cv)
}
\arguments{
\item{mp_cv}{A fitted 'mp_cv' class object returned by \code{mpersonalized_cv} function}
}
\value{
Nothing
}
\description{
This function plots the cross validation error as a function of the tuning parameters.
For penalties with 2 tuning parameters, a heat map will be plotted via the \code{image()} function
}
\examples{
set.seed(123)
sim_dat  = simulated_dataset(n = 200, problem = "meta-analysis")
Xlist = sim_dat$Xlist; Ylist = sim_dat$Ylist; Trtlist = sim_dat$Trtlist

# fit different rules with lasso penalty for this meta-analysis problem
mp_cvmod_diff = mpersonalized_cv(problem = "meta-analysis",
                                 Xlist = Xlist, Ylist = Ylist, Trtlist = Trtlist,
                                 penalty = "lasso", single_rule = FALSE)

plots = plotCV(mp_cvmod_diff)
set.seed(NULL)
}
