% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mp_cv}
\alias{predict.mp_cv}
\title{Predict Function for "mp_cv" object}
\usage{
\method{predict}{mp_cv}(mp_cv, newx = NULL, weight = NULL,
  overall_rec = TRUE)
}
\arguments{
\item{mp_cv}{the fitted "mp_cv" object}

\item{newx}{the covariate matrix of the new patients. If newx = NULL, then the prediction is made for each data set based on the rule of that study.}

\item{weight}{a weight vector for the overall recommendation. If leave as \code{NULL}, a equally weighted recommendation will be made.}

\item{overall_rec}{a logical value. If \code{TRUE}, an overall recommendation will be made weighted by the "weight" parameter.}
}
\value{
\item{treatment}{recommended treatment for each patient for each study/outcome. If \code{overall_rec = TRUE}, the weighted overall recommended treatment will be computed as well.
If the overall recommened treatment is equal to 0.5, it means the sum of weight is equal for 0 and 1.}
\item{benefit_score}{the benefit score computed from \eqn{g_1, \dots, g_K}}
}
\description{
This function predict the benefit scores and optimal treatment for new patients
}
