% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_dataset.R
\name{simulated_dataset}
\alias{simulated_dataset}
\title{Simulated Dataset Generator}
\usage{
simulated_dataset(n, sim_seed)
}
\arguments{
\item{n}{sample size}

\item{sim_seed}{the seed supplied to the generation}
}
\value{
A simulated dataset.
\item{Xlist}{a list object with \eqn{k}th element denoting the covariate matrix of study k}
\item{Ylist}{a list object with \eqn{k}th element denoting the response vector of study k}
\item{Trtlist}{a list object with \eqn{k}th element denoting the treatment vector of study k (coded as 0 or 1)}
}
\description{
This function could generate a simulated dataset for the test and usage of this package.
}
