% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_estimate.R
\name{lambda_estimate}
\alias{lambda_estimate}
\title{Default Penalty Parameter Sequence if Not Given}
\usage{
lambda_estimate(modelXlist, modelYlist, penalty, unique_rule, alpha)
}
\arguments{
\item{modelXlist}{the Xlist in contrast framework after standardization}

\item{modelYlist}{the Ylist in contrast framework after standardization}

\item{penalty}{penalty type}

\item{unique_rule}{whether a unique treatment rule is required}
}
\value{
estimated lambda for required penalty if not provided
}
