% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mp}
\alias{plot.mp}
\title{Interaction Plot for an "mp" Class Object.}
\usage{
\method{plot}{mp}(x, penalty_index, ...)
}
\arguments{
\item{x}{A fitted "mp" class object returned by \code{mpersonalzied} function}

\item{penalty_index}{The index of penalty parameter configuration in \code{mp$penalty_parameter_sequence}.
When \code{mp$penalty = "none"}, \code{penalty_index} is automatically set to be 1.}

\item{...}{not used}
}
\value{
A list object with each element as the interaction plots for a penalty parameter configuration.
}
\description{
This function plots interaction between received treatment and recommended treatment,
which provides an estimate of treatment effect of the identified subgroup.
}
\details{
In the interaction plot, each point is the group mean given a received treatment
and a recommended treatment. Although usually
overestimating treatment effect in training set, interaction plots provides a sanity check for treatment
recommendation rules. Given a specific index of penalty parameter, the function
plots corresponding interaction plots.
}
\examples{
set.seed(123)
sim_dat  = simulated_dataset(n = 200, problem = "meta-analysis")
Xlist = sim_dat$Xlist; Ylist = sim_dat$Ylist; Trtlist = sim_dat$Trtlist

# fit different rules with SGL penalty for this meta-analysis problem
mp_mod_diff = mpersonalized(problem = "meta-analysis",
                            Xlist = Xlist, Ylist = Ylist, Trtlist = Trtlist,
                            penalty = "lasso", single_rule = FALSE)

# interaction plot of the 5th penalty parameter
plots = plot(x = mp_mod_diff, penalty_index = 5)
set.seed(NULL)
}
